store: Get rid of `structs.go` as it only contains errors/config
